Update content for introduction.md
Hello @NolwennP ,
I took the liberty to propose some changes.
I'm also a bit confused on who this content is designed for, between end-users who have a DAM and a PIM account, and SI who would likely be the ones setting up the connector.
On a side note, is it necessary to have both DAM and PIM platforms up and running to configure the connection, or could it be done while projects are under development? That could be a great info in this intro!